pyproject.toml
[project]
name = "venturalitica"
version = "0.5.0"
description = "Frictionless Governance for AI. Enforce policies in your ML training with one line of code."
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Venturalítica", email = "hello@venturalitica.com"}
]
keywords = ["ai", "governance", "fairness", "mlops", "eu-ai-act", "oscal", "compliance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyYAML>=6.0.3",
    "pandas>=2.3.0",
    "numpy>=2.4.0",
    "scikit-learn>=1.8.0",
    "cyclonedx-python-lib>=11.6.0",
    "typer>=0.21.0",
    "rich>=14.2.0",
    "streamlit>=1.53.0",
    "compliance-trestle>=3.11.0",
    "ucimlrepo>=0.0.7",
    "langgraph>=1.0.7",
    "langchain-ollama>=1.0.1",
    "langchain-mistralai>=1.1.0",
    "langchain-huggingface>=0.1.0",
    "langchain-community",
    "huggingface-hub",
    "llama-cpp-python",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-macros-plugin>=1.5.0",
    "mkdocs-static-i18n>=1.3.0",
    "pymdown-extensions>=10.20.1",
    "mkdocs-llmstxt>=0.5.0",
    "posthog>=3.5.0",
]

[project.urls]
Homepage = "https://venturalitica.com"
Documentation = "https://venturalitica.github.io/venturalitica-sdk/"
Repository = "https://github.com/Venturalitica/venturalitica-sdk"
Changelog = "https://github.com/Venturalitica/venturalitica-sdk/releases"

[project.scripts]
venturalitica = "venturalitica.cli:app"

[project.optional-dependencies]
metrics = [
    "fairlearn>=0.10.0",
]
torch = [
    "torch>=2.0.0",
]
full = [
    "venturalitica[metrics,torch,green]",
]
green = [
    "codecarbon>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/venturalitica"]

[dependency-groups]
dev = [
    "clearml>=2.1.2",
    "codecarbon>=3.2.1",
    "fairlearn>=0.13.0",
    "mlflow>=3.1.4",
    "psutil>=7.2.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "wandb>=0.24.0",
]
docs = [
    "mkdocs-llmstxt>=0.4.0",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=0.30.1",
    "mkdocs-macros-plugin>=1.0.0",
    "mkdocs-static-i18n>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers -v"
filterwarnings = [
    # Suppress external library warnings
    "ignore::DeprecationWarning:clearml.utilities.pyhocon",
    "ignore::DeprecationWarning:pyparsing",
    "ignore::RuntimeWarning:numpy",
    "ignore::FutureWarning:pandas",
    # Only show warnings from our code
    "default::UserWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
# Standard rules: E (pycodestyle errors), F (pyflakes), I (isort imports)
select = ["E", "F", "I"]
# Ignore line too long (E501 - handled by format)
ignore = ["E501"]
# Exclude certain patterns
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".tox",
    ".eggs",
    "*.egg",
    "build",
    "dist",
]

[tool.ruff.lint.isort]
# isort configuration for import sorting
known-first-party = ["venturalitica"]
known-third-party = ["pandas", "numpy", "pytest"]